split update.hpp
